fix: LC-61 move pagination.ts to common folder
